requirements.txt
click
pvporcupine
pycairo
pyusb
voice-engine
fastapi
uvicorn
pydantic
websockets
pyaudio
numpy
requests